+++ types/src/main/java/com/cobber/fta/Plugins.java
@@ -178,7 +178,7 @@ public class Plugins {
 		final Object logical;
 
 		try {
-			newLogicalType = Class.forName(plugin.clazz);
+			newLogicalType = Class.forName(plugin.clazz, false, Thread.currentThread().getContextClassLoader());
 			if (!LogicalType.class.isAssignableFrom(newLogicalType))
 				throw new FTAPluginException("Semantic type: " + plugin.clazz + " does not appear to be a Semantic Type.");
 			ctor = newLogicalType.getConstructor(PluginDefinition.class);